# backend/src/main/java/com/jaee/dto/product/ProductDto.java
package com.jaee.dto.product;

import com.jaee.entity.Product;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;
import java.time.LocalDateTime;
import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ProductDto {
    private Long id;
    private String name;
    private String slug;
    private String description;
    private BigDecimal price;
    private String currency;
    private Long categoryId;
    private String categoryName;
    private List<String> images;
    private Integer stockQty;
    private Boolean active;
    private Boolean inStock;
    private LocalDateTime createdAt;
    
    public static ProductDto fromEntity(Product product) {
        return ProductDto.builder()
                .id(product.getId())
                .name(product.getName())
                .slug(product.getSlug())
                .description(product.getDescription())
                .price(product.getPrice())
                .currency(product.getCurrency())
                .categoryId(product.getCategory() != null ? product.getCategory().getId() : null)
                .categoryName(product.getCategory() != null ? product.getCategory().getName() : null)
                .images(product.getImages())
                .stockQty(product.getStockQty())
                .active(product.getActive())
                .inStock(product.isInStock())
                .createdAt(product.getCreatedAt())
                .build();
    }
}