# backend/src/main/java/com/jaee/entity/CartItem.java
package com.jaee.entity;

import jakarta.persistence.*;
import lombok.*;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Entity
@Table(name = "cart_items")
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CartItem {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "cart_id", nullable = false)
    private Cart cart;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "product_id", nullable = false)
    private Product product;

    @Column(nullable = false)
    private Integer qty;

    @Column(precision = 10, scale = 2)
    private BigDecimal unitPriceSnapshot;

    @Builder.Default
    private LocalDateTime createdAt = LocalDateTime.now();

    public BigDecimal getSubtotal() {
        return unitPriceSnapshot.multiply(BigDecimal.valueOf(qty));
    }
}